server/pyproject.toml
[project]
name = "voice-analysis-server"
version = "0.1.0"
description = "Local FastAPI backend for the Voice Analysis Web MVP."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.0",
    "python-dotenv>=1.2.0",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.35.0",
]

[tool.uv]
package = false
